ssh: do our own proxycommand handling
Adapt the code we were using for libssh2 and use it for both
libssh and libssh2 so that we have consistent proxy command
support on both unix and windows, and both versions of the
ssh libraries.

This may help resolve the linked issue on windows, or at least,
uncover the next layer of problem(s).

refs: https://github.com/wez/wezterm/issues/3083