mux: Don't add excessively long lines to the logical lines
The behavior of impl_get_logical_lines_via_get_lines() was inconsistent
when all physical lines appeared to be one wrapped logical line.

The combined logical line would be appended to the result if and only if
it exceeded MAX_LOGICAL_LINE_LEN (1024) characters.

Having an extra logical line in the result would lead to an extra newline
being added to the text copied to clipboard.

Fixes #3177.