ci: don't use release mode for tests
The original thinking was that it would be quicker to reuse
the existing release mode deps, but I'm not sure if that is
still true, so let's test it to see.

At the time of writing, the test phase takes 8-12 minutes depending
on the OS, with most of that being compilation.