ci: Use scccache-action instead of rust-cache
It reduced ci times for fluke from 9min to 1min in MacOS (see [this
issue](https://github.com/hapsoc/fluke/issues/118)).

This [comment](https://github.com/wez/wezterm/pull/4468#issuecomment-1789638340)
has the TL;DR of the results from adding it. The summary of that is
caching works on MacOS now and it uses the github cache much better.