docs: remove selection restriction
This used to make things such that the copy button for `console`
blocks would only copy the input and not the output, but it
doesn't seem to have that effect any more, and now just prevents
selecting portions of the output.

That's a PITA because eg: the CLI --help pages use `console`
to show which options are available.

I think something has probably changed somewhere in upstream mkdocs
because this used to work as I mentioned above.

refs: https://github.com/squidfunk/mkdocs-material/issues/6697