WIP fix for i3 terminals sending mouse press before focus event
I don't expect this is the correct place for this fix but the thinking
is that the window only gets the mouse press event if the mouse is over
the terminal window. In this case it makes sense (to me?) that if the
window doesn't have focus, the click should still be passed to the
application.