front_end defaults to OpenGL again
I meant to flip this back prior to cutting the most recent release,
and we've had a couple of issues already raised about WebGpu
and is transparency handling on Windows, and color/gamma.

I think I'll just leave this at OpenGL for a while, until WebGpu
catches up a bit better.