x11: scope the cyrillic ctrl-c workaround to cyrillic
The workaround was to allow ctrl-c to behave as expected for users
of cyrillic layouts, but it was scoped too broadly, as it impacted
eg: pressing ctrl-grave; in a spanish layout that is expected to
expand to º

This commit adjusts the scope of the workaround to only activate
when the original expansion is itself cyrillic.

I doubt this is the last we'll hear of this, but let's see
how it goes!

refs: https://github.com/wez/wezterm/issues/4933