nix: fix version string
Derive the version string using available information. This will work well for
clones of this repository, but not pure sources (e.g. GitHub source downloads).

Additionally, don't override the version found in existing .tag to allow actual
version to be specified manually. The side effect of this is that the actual
version might differ from that of the derivation, but there are no trivial
workarounds to set the derivation version based on existing .tag file given
that the flakes are stored in a subdirectory instead of the root of the
repository.