Fix inconsistent spawn behaviour when using default_mux_server_domain.
When using an exec domain as a default_mux_server_domain, you currently
witness the following behavior:
- Newly started mux-servers create a pane in the exec domain.
- Splitting an existing pane will create the new pane in the exec
  domain.
- Calling [wezterm cli spawn] will create a new tab or window in the
  exec domain.
- Creating a new tab from the Gui does *not* run in the exec domain.

This appears to be due to the fact that ClientDomain hardcodes an
assumption that the default domain_id of the remote mux-server is 0.

Instead of having ClientDomain issue a SpawnV2 RPC explicitly asking for
domain_id 0, simply use DefaultDomain.