shotgun fix images not being correctly removed from lines (for demonstration)
Go through all cells removing the specified image. This could have
extremely bad performance for large scrollbacks and a lot of updating
images.
The PlacementInfo should be sufficient to just remove the image from the
relevant lines, but right now this PlacementInfo can become stale after
resizing, which seems to indicate a bug.