fix: skip FreshLine by semantic integration in ble.sh
ble.sh keeps track of the cursor position in the terminal
display---including the horizontal position relative to the bottom of
the terminal---based on the terminal control sequences that it sends.
However, with WezTerm's shell integration, WezTerm inserts the newline
(FreshLine) in PROMPT_COMMAND, which breaks the cursor position
tracking of ble.sh.  The needed cursor position adjustment at the end
of the command execution is already performed by ble.sh, so WezTerm
does not need to adjust it.  In this patch, we skip the request for
FreshLine.