rudimentary attempt at fixing line wrap, part 2
This feels more like a workaround.
I would expect, that when a program prints exactly
as many chars as there a columns, the cursor would wrap
into the next line. This would be possible by putting
another "if self.wrap_next {" block after the for loop
in the performer. But it seems like there are programs
that depend on it not doing that, e.g. fish shell.
It prints a full line of mostly spaces and then does \r,
expecting the cursor is still on the same line.

This way, the cursor overflows into an extra column,
but I'm sure this has its own disadvantages. Perhaps
this commit should not be used at all.